Split OpticGraph implementation in various submodules.
