+++ opossum_core/src/nodes/node_group/optic_graph/mod.rs
@@ -1,0 +1,10 @@
+#![warn(missing_docs)]
+
+mod analysis;
+mod construction;
+mod graph;
+mod inspection;
+mod serialization;
+mod visualization;
+
+pub use graph::{ConnectionInfo, OpticGraph};
